Work-to-date on Cisco 2500 LANCE port
Still a WIP with some TODO's sprinkled throughout.

Although it "works" and I can, for a short while, ping flood with a few hundred pps, something is happening to cause the CPU to be halted and no futrher interrupts are acknowledged.

My observations to date seem to indicate that the DSACK pins of the CPU are no longer being toggled to indicate the end of a bus cycle. These signals are driven by a Cisco proprietary chip, so maybe there is another register or something that needs to be poked in there to make it work.

After spending the best part of a week debugging and trying everything, Im no longer making significant progress, so this work is being shelved for now.