Reimplemented TX and RX event handlers
Using task notifications instead of a queue, this has made a dramatic improvement.

Now able to ping flood around 1000pps and over a test period of around 10 or so minutes, 1.1 million packets were sent and received with only 2 drops and no crashes!

FreeRTOS tasks continued to run albeit VERY slowly - but that is to be expected when almost all of the processing power has been given up to packet handling.